Move UserDirectories struct in userdirs package
