+++ internal/userdirs/config_unix_test.go
@@ -1,4 +1,4 @@
-//go:build aix || darwin || dragonfly || freebsd || (js && wasm) || nacl || linux || netbsd || openbsd || solaris
+//go:build aix || dragonfly || freebsd || (js && wasm) || nacl || linux || netbsd || openbsd || solaris
 
 package userdirs_test
